Hook up application with mongoDB driver
We wire up the application with storing our books data in a mongoDB collection. We play around on how to retrieve multiple books or a single book from the database. We insert books using adminRoutes.js. We communicate with mongoDB using the native driver, that is, we are NOT using an ORM like Mongoose.