Add authentication and authorization using passport's local strategy
He set up user sign up and user sign in forms. We enable a mechanism to store users in our mongoDB. Using the passport service we fetch the users from the database using the provided username and password. If no user is returned we redirect the user to the home page. Using express middleware we can secure individual routes using all() inside of a route or we can secure a whole unit of routes by specifying an overarching bookRouter.use() statement.